+++ contrib/polygon/backfill/backfill_test.go
@@ -47,7 +47,7 @@ func (s *BackfillTests) TestTicksToBars(c *C) {
 			},
 		}
 		symbol := "AAPL"
-		exchangeIDs := []byte{9, 17}
+		exchangeIDs := []int{9, 17}
 		key := io.NewTimeBucketKeyFromString("AAPL/1Min/OHLCV")
 		model := models.NewBar(symbol, "1Min", 1440)
 
@@ -67,11 +67,11 @@ func (s *BackfillTests) TestTicksToBars(c *C) {
 		c.Assert(csm[*key].GetColumn("High").([]float64), DeepEquals, []float64{300.1})
 		c.Assert(csm[*key].GetColumn("Low").([]float64), DeepEquals, []float64{300})
 		c.Assert(csm[*key].GetColumn("Close").([]float64), DeepEquals, []float64{300.1})
-		c.Assert(csm[*key].GetColumn("Volume").([]int64), DeepEquals, []int64{180})
+		c.Assert(csm[*key].GetColumn("Volume").([]uint64), DeepEquals, []uint64{180})
 
 		model = models.NewBar(symbol, "1Min", 1440)
 		// And when we call tradesToBars with different set of exchanges
-		tradesToBars(ticks, model, []byte{8, 9})
+		tradesToBars(ticks, model, []int{8, 9})
 
 		csm = *model.BuildCsm()
 
@@ -86,7 +86,7 @@ func (s *BackfillTests) TestTicksToBars(c *C) {
 		c.Assert(csm[*key].GetColumn("High").([]float64), DeepEquals, []float64{300})
 		c.Assert(csm[*key].GetColumn("Low").([]float64), DeepEquals, []float64{299.9})
 		c.Assert(csm[*key].GetColumn("Close").([]float64), DeepEquals, []float64{299.9})
-		c.Assert(csm[*key].GetColumn("Volume").([]int64), DeepEquals, []int64{150})
+		c.Assert(csm[*key].GetColumn("Volume").([]uint64), DeepEquals, []uint64{150})
 	}
 
 	// With one condition: No update on High/Low, Volume & Close
@@ -97,12 +97,12 @@ func (s *BackfillTests) TestTicksToBars(c *C) {
 				Price:        300,
 				Size:         100,
 				Exchange:     9,
-				Conditions:   []byte{15},
+				Conditions:   []int{15},
 			},
 		}
 
 		symbol := "AAPL"
-		exchangeIDs := []byte{9}
+		exchangeIDs := []int{9}
 		key := io.NewTimeBucketKeyFromString("AAPL/1Min/OHLCV")
 		model := models.NewBar(symbol, "1Min", 1440)
 
@@ -116,7 +116,7 @@ func (s *BackfillTests) TestTicksToBars(c *C) {
 		c.Assert(csm[*key].GetColumn("High").([]float64), DeepEquals, []float64{})
 		c.Assert(csm[*key].GetColumn("Low").([]float64), DeepEquals, []float64{})
 		c.Assert(csm[*key].GetColumn("Close").([]float64), DeepEquals, []float64{})
-		c.Assert(csm[*key].GetColumn("Volume").([]int64), DeepEquals, []int64{})
+		c.Assert(csm[*key].GetColumn("Volume").([]uint64), DeepEquals, []uint64{})
 	}
 
 	// With conditions: Normal trade + No update on High/Low, Volume & Close
@@ -127,19 +127,19 @@ func (s *BackfillTests) TestTicksToBars(c *C) {
 				Price:        300,
 				Size:         100,
 				Exchange:     9,
-				Conditions:   []byte{0},
+				Conditions:   []int{0},
 			},
 			{
 				SipTimestamp: time.Date(2020, 1, 21, 9, 30, 0, 4, NY).UnixNano(),
 				Price:        305.2,
 				Size:         10,
 				Exchange:     9,
-				Conditions:   []byte{15},
+				Conditions:   []int{15},
 			},
 		}
 
 		symbol := "AAPL"
-		exchangeIDs := []byte{9, 8}
+		exchangeIDs := []int{9, 8}
 		key := io.NewTimeBucketKeyFromString("AAPL/1Min/OHLCV")
 		model := models.NewBar(symbol, "1Min", 1440)
 
@@ -153,7 +153,7 @@ func (s *BackfillTests) TestTicksToBars(c *C) {
 		c.Assert(csm[*key].GetColumn("High").([]float64), DeepEquals, []float64{300})
 		c.Assert(csm[*key].GetColumn("Low").([]float64), DeepEquals, []float64{300})
 		c.Assert(csm[*key].GetColumn("Close").([]float64), DeepEquals, []float64{300})
-		c.Assert(csm[*key].GetColumn("Volume").([]int64), DeepEquals, []int64{100})
+		c.Assert(csm[*key].GetColumn("Volume").([]uint64), DeepEquals, []uint64{100})
 	}
 
 	// With condition: Form-T, odd-lot and normal
@@ -164,40 +164,40 @@ func (s *BackfillTests) TestTicksToBars(c *C) {
 				Price:        300,
 				Size:         100,
 				Exchange:     9,
-				Conditions:   []byte{12},
+				Conditions:   []int{12},
 			},
 			{ // Should be excluded: odd-lot
 				SipTimestamp: time.Date(2020, 1, 21, 8, 30, 2, 0, NY).UnixNano(),
 				Price:        314,
 				Size:         99,
 				Exchange:     9,
-				Conditions:   []byte{12, 37},
+				Conditions:   []int{12, 37},
 			},
 			{ // Should be included
 				SipTimestamp: time.Date(2020, 1, 21, 8, 30, 2, 0, NY).UnixNano(),
 				Price:        299,
 				Size:         77,
 				Exchange:     8,
-				Conditions:   []byte{12},
+				Conditions:   []int{12},
 			},
 			{
 				SipTimestamp: time.Date(2020, 1, 21, 9, 30, 0, 4, NY).UnixNano(),
 				Price:        305.2,
 				Size:         10,
 				Exchange:     9,
-				Conditions:   []byte{14},
+				Conditions:   []int{14},
 			},
 			{
 				SipTimestamp: time.Date(2020, 1, 21, 9, 30, 1, 4, NY).UnixNano(),
 				Price:        315.2,
 				Size:         17,
 				Exchange:     8,
-				Conditions:   []byte{},
+				Conditions:   []int{},
 			},
 		}
 
 		symbol := "AAPL"
-		exchangeIDs := []byte{9, 8}
+		exchangeIDs := []int{9, 8}
 		key := io.NewTimeBucketKeyFromString("AAPL/1Min/OHLCV")
 		model := models.NewBar(symbol, "1Min", 1440)
 		// When we call tradesToBars
@@ -209,7 +209,7 @@ func (s *BackfillTests) TestTicksToBars(c *C) {
 		c.Assert(csm[*key].GetColumn("High").([]float64), DeepEquals, []float64{300, 315.2})
 		c.Assert(csm[*key].GetColumn("Low").([]float64), DeepEquals, []float64{299, 305.2})
 		c.Assert(csm[*key].GetColumn("Close").([]float64), DeepEquals, []float64{299, 315.2})
-		c.Assert(csm[*key].GetColumn("Volume").([]int64), DeepEquals, []int64{276, 27})
+		c.Assert(csm[*key].GetColumn("Volume").([]uint64), DeepEquals, []uint64{276, 27})
 
 	}
 }
